scraping-bot/requirements.txt
requests
openpyxl
beautifulsoup4
psycopg2
datetime
python-dateutil